+++ test/sync/widgets/outbox_badge_test.dart
@@ -13,7 +13,7 @@ void main() {
 
   group('OutboxBadge Widget Tests - ', () {
     setUp(() {
-      getIt.registerSingleton<ColorsService>(ColorsService(watch: false));
+      getIt.registerSingleton<ThemesService>(ThemesService(watch: false));
     });
     tearDown(getIt.reset);